backend/requirements.txt
aiofiles
cloudinary
fastapi
ffmpeg-python
firebase-admin
google-cloud-aiplatform
google-cloud-bigquery>=3.11.0
google-cloud-firestore>=2.11.0
google-cloud-storage>=2.10.0
google-generativeai>=0.8.0
google-auth>=2.22.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
gtts
hiredis>=2.2.0
instagrapi
numpy>=1.24.0
openai-whisper
pandas>=2.0.0
pillow
playsound
pydantic
pydub
python-dotenv>=1.0.1
python-multipart
redis>=4.5.0
reportlab
requests
scipy
slowapi>=0.1.8
sounddevice
SpeechRecognition
twilio
uvicorn